Add an example of connecting a Realm React app to sync
