+++ packages/realm-react/example/App.tsx
@@ -1,6 +1,6 @@
 ////////////////////////////////////////////////////////////////////////////
 //
-// Copyright 2021 Realm Inc.
+// Copyright 2022 Realm Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,114 +15,44 @@
 // limitations under the License.
 //
 ////////////////////////////////////////////////////////////////////////////
-import React, { useCallback, useMemo } from "react";
-import { SafeAreaView, View, StyleSheet } from "react-native";
 
-import TaskContext, { Task } from "./app/models/Task";
-import IntroText from "./app/components/IntroText";
-import AddTaskForm from "./app/components/AddTaskForm";
-import TaskList from "./app/components/TaskList";
-import colors from "./app/styles/colors";
+// 1. Move your App code to e.g. AppMain.tsx and create this outer App.tsx which is
+// responsible for logging in the user and setting up sync on the Realm.
 
-const { useRealm, useQuery, RealmProvider } = TaskContext;
+import React, { useEffect, useState } from "react";
+import { Realm } from "@realm/react";
 
-function App() {
-  const realm = useRealm();
-  const result = useQuery(Task);
+import TaskContext from "./app/models/Task";
+import { AppMain } from "./AppMain";
 
-  const tasks = useMemo(() => result.sorted("createdAt"), [result]);
+// 2. Add your Realm app details here, using whatever login mechanism you wish
+const app = new Realm.App({ id: "application-0-bzsvu" });
+const credentials = Realm.Credentials.anonymous();
 
-  const handleAddTask = useCallback(
-    (description: string): void => {
-      if (!description) {
-        return;
-      }
+const App = () => {
+  // 3. Store the logged in user in state
+  const [user, setUser] = useState<Realm.User | undefined>();
 
-      // Everything in the function passed to "realm.write" is a transaction and will
-      // hence succeed or fail together. A transcation is the smallest unit of transfer
-      // in Realm so we want to be mindful of how much we put into one single transaction
-      // and split them up if appropriate (more commonly seen server side). Since clients
-      // may occasionally be online during short time spans we want to increase the probability
-      // of sync participants to successfully sync everything in the transaction, otherwise
-      // no changes propagate and the transaction needs to start over when connectivity allows.
-      realm.write(() => {
-        realm.create("Task", Task.generate(description));
-      });
-    },
-    [realm],
-  );
+  const { RealmProvider } = TaskContext;
 
-  const handleToggleTaskStatus = useCallback(
-    (task: Task): void => {
-      realm.write(() => {
-        // Normally when updating a record in a NoSQL or SQL database, we have to type
-        // a statement that will later be interpreted and used as instructions for how
-        // to update the record. But in RealmDB, the objects are "live" because they are
-        // actually referencing the object's location in memory on the device (memory mapping).
-        // So rather than typing a statement, we modify the object directly by changing
-        // the property values. If the changes adhere to the schema, Realm will accept
-        // this new version of the object and wherever this object is being referenced
-        // locally will also see the changes "live".
-        task.isComplete = !task.isComplete;
-      });
+  // 4. On application startup, login the user
+  useEffect(() => {
+    (async () => {
+      const loggedInUser = await app.logIn(credentials);
+      // 5. Set the logged in user in state
+      setUser(loggedInUser);
+    })();
+  }, []);
 
-      // Alternatively if passing the ID as the argument to handleToggleTaskStatus:
-      // realm?.write(() => {
-      //   const task = realm?.objectForPrimaryKey('Task', id); // If the ID is passed as an ObjectId
-      //   const task = realm?.objectForPrimaryKey('Task', Realm.BSON.ObjectId(id));  // If the ID is passed as a string
-      //   task.isComplete = !task.isComplete;
-      // });
-    },
-    [realm],
-  );
+  // 6. If we are not logged in yet, return null so the app does not render without sync enabled
+  if (!user) return null;
 
-  const handleDeleteTask = useCallback(
-    (task: Task): void => {
-      realm.write(() => {
-        realm.delete(task);
-
-        // Alternatively if passing the ID as the argument to handleDeleteTask:
-        // realm?.delete(realm?.objectForPrimaryKey('Task', id));
-      });
-    },
-    [realm],
-  );
-
+  // 7. If we are logged in, add the sync configuration the the Realm and render the ap
   return (
-    <SafeAreaView style={styles.screen}>
-      <View style={styles.content}>
-        <AddTaskForm onSubmit={handleAddTask} />
-        {tasks.length === 0 ? (
-          <IntroText />
-        ) : (
-          <TaskList tasks={tasks} onToggleTaskStatus={handleToggleTaskStatus} onDeleteTask={handleDeleteTask} />
-        )}
-      </View>
-    </SafeAreaView>
-  );
-}
-
-const styles = StyleSheet.create({
-  screen: {
-    flex: 1,
-    backgroundColor: colors.darkBlue,
-  },
-  content: {
-    flex: 1,
-    paddingTop: 20,
-    paddingHorizontal: 20,
-  },
-});
-
-function AppWrapper() {
-  if (!RealmProvider) {
-    return null;
-  }
-  return (
-    <RealmProvider>
-      <App />
+    <RealmProvider sync={{ user, partitionValue: "default" }}>
+      <AppMain />
     </RealmProvider>
   );
-}
+};
 
-export default AppWrapper;
+export default App;
