+++ packages/realm-react/example/__tests__/App-test.tsx
@@ -18,7 +18,7 @@
 
 import "react-native";
 import React from "react";
-import App from "../App";
+import App from "../AppMain";
 
 // Note: test renderer must be required after react-native.
 import renderer from "react-test-renderer";
